cmd: framwork: Remove impl_x from cmd api
