+++ src/include/framework/cmd/api.h
@@ -9,10 +9,32 @@
 #ifndef FRAMEWORK_CMD_API_H_
 #define FRAMEWORK_CMD_API_H_
 
-#include __impl_x(framework/cmd/api_impl.h)
+#include <util/array.h>
+#include <framework/mod/api.h>
+
+#include <framework/cmd/types.h>
 
 #define cmd_foreach(cmd) \
-		__cmd_foreach(cmd)
+	array_spread_foreach(cmd, __cmd_registry)
+
+ARRAY_SPREAD_DECLARE(const struct cmd *, __cmd_registry);
+
+#define CMD_ADD(_cmd_ptr) \
+	ARRAY_SPREAD_DECLARE(const struct cmd *, __cmd_registry); \
+	ARRAY_SPREAD_ADD(__cmd_registry, _cmd_ptr)
+
+static inline const char *cmd_name(const struct cmd *cmd) {
+	return cmd->desc ? cmd->desc->name : NULL;
+}
+
+static inline const char *cmd_brief(const struct cmd *cmd) {
+	return cmd->desc ? cmd->desc->brief : NULL;
+}
+
+static inline const char *cmd_details(const struct cmd *cmd) {
+	return cmd->desc ? cmd->desc->details : NULL;
+}
+
 
 /**
  * TODO docs. -- Eldar
@@ -22,11 +44,12 @@ struct cmd;
 extern int cmd_exec(const struct cmd *cmd, int argc, char **argv);
 
 extern const struct cmd *cmd_lookup(const char *name);
-
+#if 0
 extern const char *cmd_name(const struct cmd *cmd);
 
 extern const char *cmd_brief(const struct cmd *cmd);
 
 extern const char *cmd_details(const struct cmd *cmd);
+#endif
 
 #endif /* FRAMEWORK_CMD_API_H_ */
